requirements.txt
flask==2.0.1
pandas==1.3.3
python-docx==0.8.11
mysql-connector-python==8.0.26
openpyxl==3.0.7